requirements.txt
openai==0.27.6
python-dotenv
fastapi
uvicorn
